+++ extensions/amazon-aws/src/components/codepipeline/stop-execution-action.tsx
@@ -6,17 +6,17 @@ import {
   StopPipelineExecutionCommand,
 } from "@aws-sdk/client-codepipeline";
 import { getErrorMessage } from "../../util";
+import { Pipeline } from "../../codepipeline";
+import { MutatePromise } from "@raycast/utils";
 
 export const StopExecutionAction = ({
-  pipelineName,
-  revalidatePipeline,
+  pipeline,
   isLoading,
-  executions,
+  mutate,
 }: {
-  pipelineName: string;
-  executions?: PipelineExecutionSummary[];
+  pipeline: Pipeline;
   isLoading: boolean;
-  revalidatePipeline: () => void;
+  mutate: MutatePromise<Pipeline[]>;
 }) => {
   return (
     <ActionPanel.Submenu
@@ -26,14 +26,14 @@ export const StopExecutionAction = ({
       isLoading={isLoading}
       filtering
     >
-      {(executions ?? [])
+      {pipeline.executions
         .filter((e) => e.status === PipelineExecutionStatus.InProgress)
         .map((e) => (
           <Action
             key={`${e.pipelineExecutionId}`}
             icon={{ source: Icon.Hourglass, tintColor: Color.Blue }}
             title={`${e.pipelineExecutionId}`}
-            onAction={() => stopExecution(pipelineName, e, revalidatePipeline)}
+            onAction={() => stopExecution(pipeline.name!, e, mutate)}
           />
         ))}
     </ActionPanel.Submenu>
@@ -42,26 +42,26 @@ export const StopExecutionAction = ({
 
 const stopExecution = async (
   pipelineName: string,
-  e: { pipelineExecutionId?: string },
-  revalidatePipeline: () => void,
+  execution: PipelineExecutionSummary,
+  mutate: MutatePromise<Pipeline[]>,
 ) => {
   await confirmAlert({
     title: "Are you sure?",
     icon: { source: Icon.Stop, tintColor: Color.Red },
-    message: `Stop ${e.pipelineExecutionId} for ${pipelineName}?`,
+    message: `Stop ${execution.pipelineExecutionId} for ${pipelineName}?`,
     primaryAction: {
       title: "Stop",
       onAction: async () => {
         const toast = await showToast(
           Toast.Style.Animated,
           `❗Stopping execution for ${pipelineName}`,
-          `Execution ID: ${e.pipelineExecutionId}`,
+          `Execution ID: ${execution.pipelineExecutionId}`,
         );
         new CodePipelineClient({})
           .send(
             new StopPipelineExecutionCommand({
               pipelineName,
-              pipelineExecutionId: e.pipelineExecutionId,
+              pipelineExecutionId: execution.pipelineExecutionId,
               abandon: true,
               reason: "Abandoned by Raycast",
             }),
@@ -77,7 +77,7 @@ const stopExecution = async (
             toast.title = `❌ Failed to stop execution for ${pipelineName}`;
             toast.message = getErrorMessage(err);
           })
-          .finally(revalidatePipeline);
+          .finally(mutate);
       },
     },
   });